move analysis reset fncs into formsUtils
