+++ analysis/js/index.js
@@ -2,9 +2,9 @@ import makeMap from './map/map.js'
 import sources from './map/mapSources.js'
 import layers from './map/mapLayers.js'
 import mapUtils from './map/mapUtils.js'
+import handleForms from './sidebar/forms.js'
 import { makePopup, makePopupContent } from './map/popup.js'
-import { handleForms, resetAnalysisLayers } from './sidebar/forms.js'
-import { resetLTSLayers } from './sidebar/formsUtils.js'
+import { resetLTSLayers, resetAnalysisLayers } from './sidebar/formsUtils.js'
 // import createFeedbackForm from './sidebar/feedback.js'
 
 const sidebar = document.getElementById('sidebar')
