+++ analysis/js/sidebar/formsUtils.js
@@ -1,4 +1,5 @@
-import { handleLegend } from "./legends.js"
+import { handleLegend, clearAnalysisLegends } from "./legends.js"
+import { analysisLookup } from "./formsConfigs.js"
 
 const resetLTSLayers = (map, e) => {
     const btn = e.target
@@ -16,4 +17,15 @@ const resetLTSLayers = (map, e) => {
     handleLegend('lts', false, 4)
 }
 
-export { resetLTSLayers }
+const resetAnalysisLayers = map => {
+    clearAnalysisLayers(map)
+    clearAnalysisLegends()
+}
+
+const clearAnalysisLayers = map => {
+    analysisLookup.forEach(layer => {
+        if(map.getLayer(layer)) map.setLayoutProperty(layer, 'visibility', 'none')
+    })
+}
+
+export { resetLTSLayers, resetAnalysisLayers, clearAnalysisLayers }
